Add Pydantic schemas for feedback, memory, optimization, runtime, sandbox, sessions, volumes, and websocket APIs
- Implemented `TraceFeedbackRequest` and `TraceFeedbackResponse` for MLflow trace feedback handling.
- Created `MemoryItemResponse` and `MemoryListResponse` for memory browsing endpoints.
- Developed comprehensive schemas for GEPA optimization requests and responses, including validation logic.
- Added runtime settings schemas for managing runtime configurations and connectivity tests.
- Introduced sandbox-related schemas for managing sandbox instances and their details.
- Established session management schemas for handling chat sessions, including state summaries and turn details.
- Defined volume management schemas for interacting with persistent storage volumes.
- Implemented websocket message schemas for chat and command interactions in the FastAPI server.
