+++ src/fleet_rlm/api/schemas/core.py
@@ -1,1087 +1,27 @@
-"""Pydantic request/response schemas for the FastAPI server."""
+"""Backward-compatible schema re-exports."""
 
 from __future__ import annotations
 
-from typing import Any, Literal
-
-from pydantic import BaseModel, Field, model_validator
-from pydantic_core import PydanticCustomError
-
-from fleet_rlm import __version__
-
-ExecutionMode = Literal["auto", "rlm_only", "tools_only"]
-VolumeProvider = Literal["daytona"]
-
-
-class HealthResponse(BaseModel):
-    """Response body for the lightweight health endpoint."""
-
-    ok: bool = Field(
-        default=True, description="Whether the service reports itself as healthy."
-    )
-    version: str = Field(
-        default=__version__,
-        description="Package version currently serving the API.",
-    )
-
-
-class ReadyResponse(BaseModel):
-    """Response body for the readiness endpoint."""
-
-    ready: bool = Field(description="Whether critical startup dependencies are ready.")
-    planner_configured: bool = Field(
-        description="Whether a planner model is currently configured and available."
-    )
-    planner: Literal["ready", "missing"] = Field(
-        description="Planner readiness classification."
-    )
-    database: Literal["ready", "missing", "disabled", "degraded"] = Field(
-        description="Database readiness classification for persistence-backed features."
-    )
-    database_required: bool = Field(
-        description="Whether the current server configuration requires database availability."
-    )
-    sandbox_provider: str = Field(
-        description="Active sandbox backend selected for runtime execution."
-    )
-
-
-class AuthMeResponse(BaseModel):
-    """Resolved identity payload returned to authenticated clients."""
-
-    tenant_claim: str = Field(
-        description="Tenant or workspace claim resolved from auth."
-    )
-    user_claim: str = Field(description="User claim resolved from auth.")
-    email: str | None = Field(
-        default=None,
-        description="User email address when the auth provider returned one.",
-    )
-    name: str | None = Field(
-        default=None,
-        description="Display name returned by the auth provider, when available.",
-    )
-    tenant_id: str | None = Field(
-        default=None,
-        description="Persisted control-plane tenant identifier for admitted Entra users.",
-    )
-    user_id: str | None = Field(
-        default=None,
-        description="Persisted control-plane user identifier for admitted Entra users.",
-    )
-
-
-class WSMessage(BaseModel):
-    """Typed websocket payload for chat, cancel, and command frames."""
-
-    type: Literal["message", "cancel", "command"] = Field(
-        default="message",
-        description="Websocket frame type.",
-    )
-    content: str = Field(
-        default="", description="Primary chat content for message frames."
-    )
-    docs_path: str | None = Field(
-        default=None,
-        description="Optional local documentation path to preload before execution.",
-    )
-    trace: bool = Field(
-        default=True,
-        description="Whether trace-oriented streaming events should be emitted for the turn.",
-    )
-    trace_mode: Literal["compact", "verbose", "off"] | None = Field(
-        default=None,
-        description="Optional trace verbosity override for the websocket session.",
-    )
-    execution_mode: ExecutionMode = Field(
-        default="auto",
-        description="Per-turn execution mode hint for the Daytona-backed websocket runtime.",
-    )
-    repo_url: str | None = Field(
-        default=None,
-        description="Repository URL to attach to Daytona pilot runs.",
-    )
-    repo_ref: str | None = Field(
-        default=None,
-        description="Optional branch, tag, or commit to checkout for Daytona pilot runs.",
-    )
-    context_paths: list[str] | None = Field(
-        default=None,
-        description="Optional repository paths to prioritize as context for Daytona pilot runs.",
-    )
-    batch_concurrency: int | None = Field(
-        default=None,
-        description="Optional Daytona concurrency hint for batched repository work.",
-    )
-    session_id: str | None = Field(
-        default=None,
-        description="Optional session identifier for restoring an existing websocket session.",
-    )
-    # Command dispatch fields (used when type == "command")
-    command: str = Field(
-        default="", description="Command name when `type` is `command`."
-    )
-    args: dict[str, Any] = Field(
-        default_factory=dict,
-        description="Command arguments when `type` is `command`.",
-    )
-
-    @model_validator(mode="before")
-    @classmethod
-    def _validate_daytona_message_contract(cls, raw: Any) -> Any:
-        if not isinstance(raw, dict):
-            return raw
-
-        if "workspace_id" in raw or "user_id" in raw:
-            raise PydanticCustomError(
-                "unsupported_identity_fields",
-                "WebSocket identity is derived from auth. Remove workspace_id/user_id and use session_id only.",
-            )
-
-        message_type = str(raw.get("type", "message") or "message").strip()
-        if message_type == "message" and raw.get("max_depth") is not None:
-            raise PydanticCustomError(
-                "daytona_max_depth_removed",
-                "Daytona websocket requests no longer accept max_depth; use the "
-                "server-configured recursion depth.",
-            )
-
-        if (
-            message_type == "message"
-            and str(raw.get("repo_ref", "") or "").strip()
-            and not str(raw.get("repo_url", "") or "").strip()
-        ):
-            raise PydanticCustomError(
-                "daytona_repo_ref_requires_repo",
-                "Daytona repo_ref requires repo_url.",
-            )
-
-        return raw
-
-
-class WSCommandMessage(BaseModel):
-    """Typed model for a WebSocket command message."""
-
-    type: Literal["command"] = "command"
-    command: str
-    args: dict[str, Any] = Field(default_factory=dict)
-    session_id: str | None = None
-
-
-class WSCommandResult(BaseModel):
-    """Server response for a command dispatch."""
-
-    type: Literal["command_result"] = "command_result"
-    command: str
-    result: dict[str, Any] = Field(default_factory=dict)
-
-
-class SessionStateSummary(BaseModel):
-    """Lightweight summary of a persisted or active chat session."""
-
-    key: str = Field(description="Stable in-memory session key used by the server.")
-    workspace_id: str = Field(description="Workspace identifier owning the session.")
-    user_id: str = Field(description="User identifier owning the session.")
-    session_id: str | None = Field(
-        default=None,
-        description="Optional explicit session identifier when one has been assigned.",
-    )
-    history_turns: int = Field(
-        default=0,
-        description="Number of conversation turns currently stored in session history.",
-    )
-    document_count: int = Field(
-        default=0,
-        description="Number of loaded document entries attached to the session state.",
-    )
-    memory_count: int = Field(
-        default=0,
-        description="Number of persisted memory items in the session manifest.",
-    )
-    log_count: int = Field(
-        default=0,
-        description="Number of execution log entries in the session manifest.",
-    )
-    artifact_count: int = Field(
-        default=0,
-        description="Number of artifacts currently tracked in the session manifest.",
-    )
-    updated_at: str | None = Field(
-        default=None,
-        description="Last updated timestamp recorded in the session manifest, when available.",
-    )
-
-
-class SessionStateResponse(BaseModel):
-    """Response body for the session-state summary endpoint."""
-
-    ok: bool = Field(
-        default=True,
-        description="Whether the session-state query completed successfully.",
-    )
-    sessions: list[SessionStateSummary] = Field(
-        default_factory=list,
-        description="Active or restored session summaries currently known to the server.",
-    )
-
-
-# ---------------------------------------------------------------------------
-# Session history (durable transcript store)
-# ---------------------------------------------------------------------------
-
-
-class SessionListItem(BaseModel):
-    """Lightweight session summary for list views."""
-
-    id: str = Field(description="Durable session identifier.")
-    title: str = Field(description="Human-readable session title.")
-    status: str = Field(description="Session status (active, archived).")
-    model_name: str | None = Field(default=None, description="Model used in session.")
-    external_session_id: str | None = Field(
-        default=None, description="Canonical runtime session identifier."
-    )
-    created_at: str = Field(description="ISO-8601 creation timestamp.")
-    updated_at: str = Field(description="ISO-8601 last-update timestamp.")
-
-
-class SessionListResponse(BaseModel):
-    """Paginated session list."""
-
-    items: list[SessionListItem] = Field(description="Session list items.")
-    total: int = Field(description="Total matching sessions.")
-    offset: int = Field(description="Current pagination offset.")
-    limit: int = Field(description="Current page size.")
-    has_more: bool = Field(description="Whether more results exist beyond this page.")
-
-
-class SessionDetailResponse(BaseModel):
-    """Full session detail with turn count."""
-
-    id: str = Field(description="Durable session identifier.")
-    title: str = Field(description="Human-readable session title.")
-    status: str = Field(description="Session status (active, archived).")
-    model_name: str | None = Field(default=None, description="Model used in session.")
-    external_session_id: str | None = Field(
-        default=None, description="Canonical runtime session identifier."
-    )
-    workspace_id: str | None = Field(default=None, description="Workspace context.")
-    turn_count: int = Field(description="Total number of turns in this session.")
-    created_at: str = Field(description="ISO-8601 creation timestamp.")
-    updated_at: str = Field(description="ISO-8601 last-update timestamp.")
-
-
-class TurnItem(BaseModel):
-    """Single turn in a session transcript."""
-
-    id: str = Field(description="Durable turn identifier.")
-    turn_index: int = Field(description="Zero-based turn position.")
-    user_message: str = Field(description="User message text.")
-    assistant_message: str | None = Field(
-        default=None, description="Assistant response text."
-    )
-    created_at: str = Field(description="ISO-8601 creation timestamp.")
-
-
-class TurnListResponse(BaseModel):
-    """Paginated turn list."""
-
-    items: list[TurnItem] = Field(description="Turn list items.")
-    total: int = Field(description="Total turns in session.")
-    offset: int = Field(description="Current pagination offset.")
-    limit: int = Field(description="Current page size.")
-    has_more: bool = Field(description="Whether more turns exist beyond this page.")
-
-
-class SessionDeleteResponse(BaseModel):
-    """Result payload after archiving a session."""
-
-    ok: bool = Field(
-        default=True,
-        description="Whether the session was archived successfully.",
-    )
-
-
-class SessionRestoreResponse(BaseModel):
-    """Result payload after restoring an archived session."""
-
-    ok: bool = Field(
-        default=True,
-        description="Whether the session was restored successfully.",
-    )
-
-
-class SessionPatchRequest(BaseModel):
-    """Patch body for updating session metadata."""
-
-    title: str | None = Field(
-        default=None,
-        description="New human-readable session title.",
-    )
-    metadata_json: dict[str, Any] | None = Field(
-        default=None,
-        description="New metadata dictionary to merge or replace session metadata.",
-    )
-
-
-class SessionStatsResponse(BaseModel):
-    """Aggregated usage stats for a session."""
-
-    total_tokens_in: int = Field(
-        default=0,
-        description="Total input tokens across all turns.",
-    )
-    total_tokens_out: int = Field(
-        default=0,
-        description="Total output tokens across all turns.",
-    )
-    total_latency_ms: int = Field(
-        default=0,
-        description="Total latency in milliseconds across all turns.",
-    )
-    model_breakdown: dict[str, int] = Field(
-        default_factory=dict,
-        description="Mapping of model_name to turn count.",
-    )
-
-
-class RuntimeSettingsSnapshot(BaseModel):
-    """Current runtime settings snapshot returned by the Settings API."""
-
-    env_path: str = Field(
-        description="Filesystem path to the environment file being edited."
-    )
-    keys: list[str] = Field(
-        default_factory=list,
-        description="Ordered list of runtime setting keys surfaced by the Settings API.",
-    )
-    values: dict[str, str] = Field(
-        default_factory=dict,
-        description="Unmasked runtime setting values that are safe to return directly.",
-    )
-    masked_values: dict[str, str] = Field(
-        default_factory=dict,
-        description="Masked secret values returned for display-only settings fields.",
-    )
-
-
-class RuntimeSettingsUpdateRequest(BaseModel):
-    """Patch body for runtime setting updates."""
-
-    updates: dict[str, Any] = Field(
-        default_factory=dict,
-        description="Mapping of allowlisted runtime setting keys to their new values.",
-    )
-
-
-class RuntimeSettingsUpdateResponse(BaseModel):
-    """Result payload after runtime settings are persisted and hot-applied."""
-
-    updated: list[str] = Field(
-        default_factory=list,
-        description="Runtime setting keys that were successfully updated.",
-    )
-    env_path: str = Field(
-        description="Filesystem path to the environment file that was updated."
-    )
-
-
-class RuntimeConnectivityTestResponse(BaseModel):
-    """Result payload for runtime connectivity and preflight diagnostics."""
-
-    kind: Literal["lm", "daytona"] = Field(
-        description="Runtime subsystem that was tested."
-    )
-    ok: bool = Field(
-        description="Whether the connectivity test completed successfully."
-    )
-    preflight_ok: bool = Field(
-        description="Whether prerequisite configuration checks passed."
-    )
-    checked_at: str = Field(description="UTC timestamp when the test completed.")
-    checks: dict[str, Any] = Field(
-        default_factory=dict,
-        description="Structured boolean or value checks collected during the test run.",
-    )
-    guidance: list[str] = Field(
-        default_factory=list,
-        description="Human-readable remediation steps when the test did not pass cleanly.",
-    )
-    latency_ms: int | None = Field(
-        default=None,
-        description="Observed latency for the successful smoke test, when applicable.",
-    )
-    output_preview: str | None = Field(
-        default=None,
-        description="Short preview of the smoke-test output, when available.",
-    )
-    error: str | None = Field(
-        default=None,
-        description="Error summary when the test failed.",
-    )
-
-
-class RuntimeTestCache(BaseModel):
-    """Cached runtime test results included in the runtime status payload."""
-
-    lm: RuntimeConnectivityTestResponse | None = Field(
-        default=None,
-        description="Most recent language-model connectivity test result, if one has been run.",
-    )
-    daytona: RuntimeConnectivityTestResponse | None = Field(
-        default=None,
-        description="Most recent Daytona connectivity test result, if one has been run.",
-    )
-
-
-class RuntimeActiveModels(BaseModel):
-    """Resolved active model identifiers currently loaded by the runtime."""
-
-    planner: str = Field(
-        default="", description="Planner model identifier currently in use."
-    )
-    delegate: str = Field(
-        default="", description="Delegate model identifier currently in use."
-    )
-    delegate_small: str = Field(
-        default="",
-        description="Small delegate model identifier currently in use, when configured.",
-    )
-
-
-class RuntimeStatusResponse(BaseModel):
-    """Combined readiness and diagnostics snapshot for the runtime settings UI."""
-
-    app_env: str = Field(
-        description="Current application environment, such as `local` or `prod`."
-    )
-    write_enabled: bool = Field(
-        description="Whether runtime settings writes are currently allowed."
-    )
-    ready: bool = Field(
-        description="Whether critical runtime services are ready to serve requests."
-    )
-    active_models: RuntimeActiveModels = Field(
-        description="Resolved planner and delegate model identities."
-    )
-    sandbox_provider: VolumeProvider = Field(
-        default="daytona",
-        description="Active sandbox backend selected for runtime execution and volume browsing.",
-    )
-    llm: dict[str, Any] = Field(
-        default_factory=dict,
-        description="Language-model configuration and readiness diagnostics.",
-    )
-    mlflow: dict[str, Any] = Field(
-        default_factory=dict,
-        description="MLflow enablement and startup diagnostics.",
-    )
-    daytona: dict[str, Any] = Field(
-        default_factory=dict,
-        description="Daytona configuration and readiness diagnostics.",
-    )
-    tests: RuntimeTestCache = Field(
-        description="Cached runtime connectivity test results exposed in the Settings UI."
-    )
-    guidance: list[str] = Field(
-        default_factory=list,
-        description="Human-readable remediation steps for incomplete runtime setup.",
-    )
-
-
-class VolumeTreeNode(BaseModel):
-    """A single node in the volume file tree."""
-
-    id: str = Field(
-        description="Stable node identifier used by the frontend tree view."
-    )
-    name: str = Field(description="Display name for the file-system node.")
-    path: str = Field(
-        description="Absolute path for the file-system node within the runtime volume."
-    )
-    type: Literal["volume", "directory", "file"] = Field(
-        description="Kind of file-system node represented by this entry."
-    )
-    children: list[VolumeTreeNode] = Field(
-        default_factory=list,
-        description="Child nodes for directory or volume entries.",
-    )
-    size: int | None = Field(
-        default=None,
-        description="File size in bytes when the provider reports one.",
-    )
-    modified_at: str | None = Field(
-        default=None,
-        description="Last modified timestamp when the provider reports one.",
-    )
-
-
-class VolumeTreeResponse(BaseModel):
-    """Response for the volume tree listing endpoint."""
-
-    provider: VolumeProvider = Field(
-        description="Runtime volume backend used to satisfy the request."
-    )
-    volume_name: str = Field(
-        description="Resolved volume name used for the listing request."
-    )
-    root_path: str = Field(
-        description="Normalized root path used for the listing request."
-    )
-    nodes: list[VolumeTreeNode] = Field(
-        description="Tree nodes rooted at the requested path."
-    )
-    total_files: int = Field(
-        default=0,
-        description="Total file count returned in the current response payload.",
-    )
-    total_dirs: int = Field(
-        default=0,
-        description="Total directory count returned in the current response payload.",
-    )
-    truncated: bool = Field(
-        default=False,
-        description="Whether the provider truncated the tree because of depth or payload limits.",
-    )
-
-
-class VolumeFileContentResponse(BaseModel):
-    """Response for runtime volume file-content preview endpoint."""
-
-    provider: VolumeProvider = Field(
-        description="Runtime volume backend used to satisfy the request."
-    )
-    path: str = Field(description="Normalized file path used for the preview request.")
-    mime: str = Field(description="Detected MIME type for the returned content.")
-    size: int = Field(description="File size in bytes reported by the provider.")
-    content: str = Field(
-        description="UTF-8 text preview returned for the requested file."
-    )
-    truncated: bool = Field(
-        default=False,
-        description="Whether the returned file content was truncated to respect max_bytes.",
-    )
-
-
-class VolumeListItem(BaseModel):
-    """Single volume entry returned by the volume list endpoint."""
-
-    id: str = Field(description="Volume identifier.")
-    name: str = Field(description="Volume name.")
-    state: str = Field(default="", description="Volume state (e.g. ready, creating).")
-    created_at: str | None = Field(
-        default=None, description="ISO-8601 creation timestamp when available."
-    )
-
-
-class VolumeListResponse(BaseModel):
-    """Response for the volume list endpoint."""
-
-    provider: VolumeProvider = Field(
-        description="Runtime volume backend used to satisfy the request."
-    )
-    volumes: list[VolumeListItem] = Field(
-        default_factory=list, description="Available persistent volumes."
-    )
-
-
-class GEPAOptimizationRequest(BaseModel):
-    """Request body for triggering a GEPA prompt optimization run."""
-
-    dataset_path: str | None = Field(
-        default=None,
-        description="Relative filesystem path to the dataset file.",
-    )
-    dataset_id: str | None = Field(
-        default=None,
-        description="Registered dataset identifier to optimize against.",
-    )
-    program_spec: str = Field(
-        default="",
-        description="DSPy program specification string to optimize in module:attr form. "
-        "Required when module_slug is not provided.",
-    )
-    module_slug: str | None = Field(
-        default=None,
-        description="Registered module slug for server-side dispatch. "
-        "When provided, program_spec is auto-resolved from the module registry.",
-    )
-    output_path: str | None = Field(
-        default=None,
-        description="Optional filesystem path to save the optimized program.",
-    )
-    auto: Literal["light", "medium", "heavy"] = Field(
-        default="light",
-        description="GEPA optimization intensity level.",
-    )
-    train_ratio: float = Field(
-        default=0.8,
-        description="Fraction of examples to use for training (remainder used for validation).",
-    )
-
-    @model_validator(mode="after")
-    def validate_dataset_target(self) -> GEPAOptimizationRequest:
-        if self.dataset_id is not None or (self.dataset_path or "").strip():
-            return self
-        raise ValueError("dataset_id or dataset_path is required")
-
-
-class GEPAOptimizationResponse(BaseModel):
-    """Result payload after a GEPA optimization run completes."""
-
-    ok: bool = Field(
-        default=True,
-        description="Whether the optimization run completed successfully.",
-    )
-    optimizer: str = Field(
-        default="GEPA",
-        description="Optimizer backend that was used.",
-    )
-    program_spec: str = Field(
-        description="DSPy program specification that was optimized.",
-    )
-    train_examples: int = Field(
-        description="Number of training examples used.",
-    )
-    validation_examples: int = Field(
-        description="Number of validation examples used.",
-    )
-    validation_score: float | None = Field(
-        default=None,
-        description="Validation score from the optimized program, when available.",
-    )
-    output_path: str | None = Field(
-        default=None,
-        description="Filesystem path where the optimized program was saved.",
-    )
-    manifest_path: str | None = Field(
-        default=None,
-        description="Filesystem path to the optimization manifest, when available.",
-    )
-    module_slug: str | None = Field(
-        default=None,
-        description="Module slug used for this optimization run, when server-side dispatch was used.",
-    )
-    error: str | None = Field(
-        default=None,
-        description="Error message when the optimization run failed.",
-    )
-
-
-class GEPAModuleInfo(BaseModel):
-    """Metadata for a registered optimizable DSPy module."""
-
-    slug: str = Field(description="Unique module identifier slug.")
-    label: str = Field(description="Human-readable module label.")
-    description: str = Field(
-        default="",
-        description="Human-readable description of what this module optimizes.",
-    )
-    program_spec: str = Field(description="DSPy program specification string.")
-    required_dataset_keys: list[str] = Field(
-        description="Dataset keys required for this module's examples."
-    )
-
-
-class GEPAStatusResponse(BaseModel):
-    """Status payload for GEPA optimization availability."""
-
-    available: bool = Field(
-        description="Whether GEPA optimization is available in this environment.",
-    )
-    mlflow_enabled: bool = Field(
-        description="Whether MLflow is enabled and reachable.",
-    )
-    gepa_installed: bool = Field(
-        description="Whether the GEPA teleprompt module is importable.",
-    )
-    guidance: list[str] = Field(
-        default_factory=list,
-        description="Human-readable guidance when GEPA is not fully available.",
-    )
-
-
-class OptimizationRunResponse(BaseModel):
-    """A single optimization run record."""
-
-    id: str = Field(description="Unique run identifier.")
-    status: str = Field(description="Run status: running, completed, or failed.")
-    module_slug: str | None = Field(
-        default=None, description="Module slug when server-side dispatch was used."
-    )
-    program_spec: str = Field(
-        description="DSPy program specification that was optimized."
-    )
-    optimizer: str = Field(description="Optimizer backend that was used.")
-    auto: str | None = Field(
-        default="light", description="Optimization intensity level."
-    )
-    train_ratio: float = Field(default=0.8, description="Train/validation split ratio.")
-    dataset_path: str | None = Field(
-        default=None, description="Path to the dataset used."
-    )
-    train_examples: int | None = Field(
-        default=None, description="Number of training examples used."
-    )
-    validation_examples: int | None = Field(
-        default=None, description="Number of validation examples used."
-    )
-    validation_score: float | None = Field(
-        default=None, description="Validation score from the optimized program."
-    )
-    output_path: str | None = Field(
-        default=None,
-        description="Filesystem path where the optimized program was saved.",
-    )
-    manifest_path: str | None = Field(
-        default=None, description="Filesystem path to the optimization manifest."
-    )
-    error: str | None = Field(
-        default=None, description="Error message when the run failed."
-    )
-    phase: str | None = Field(
-        default=None, description="Current phase of the optimization run."
-    )
-    started_at: str = Field(description="ISO timestamp when the run started.")
-    completed_at: str | None = Field(
-        default=None, description="ISO timestamp when the run completed."
-    )
-
-
-class OptimizationRunCreatedResponse(BaseModel):
-    """Response when an async optimization run is created."""
-
-    run_id: str = Field(description="Unique identifier for the created run.")
-    status: str = Field(default="running", description="Initial run status.")
-
-
-class TraceFeedbackRequest(BaseModel):
-    """Feedback payload for annotating an MLflow trace."""
-
-    trace_id: str | None = Field(
-        default=None,
-        description="Resolved MLflow trace identifier when the client already knows it.",
-    )
-    client_request_id: str | None = Field(
-        default=None,
-        description="Client request identifier used to resolve the trace when trace_id is absent.",
-    )
-    is_correct: bool = Field(
-        description="Whether the model output was considered correct."
-    )
-    comment: str | None = Field(
-        default=None,
-        description="Optional free-form reviewer comment explaining the feedback.",
-    )
-    expected_response: str | None = Field(
-        default=None,
-        description="Optional ground-truth response or correction to log alongside the feedback.",
-    )
-
-    @model_validator(mode="after")
-    def validate_trace_lookup_target(self) -> TraceFeedbackRequest:
-        if (self.trace_id or "").strip() or (self.client_request_id or "").strip():
-            return self
-        raise ValueError("trace_id or client_request_id is required")
-
-
-class TraceFeedbackResponse(BaseModel):
-    """Result payload after MLflow feedback has been recorded."""
-
-    ok: bool = Field(
-        default=True, description="Whether the feedback request completed successfully."
-    )
-    trace_id: str = Field(
-        description="Resolved MLflow trace identifier that received the feedback."
-    )
-    client_request_id: str | None = Field(
-        default=None,
-        description="Resolved client request identifier associated with the trace, when available.",
-    )
-    feedback_logged: bool = Field(
-        default=True,
-        description="Whether binary/correctness feedback was successfully logged.",
-    )
-    expectation_logged: bool = Field(
-        default=False,
-        description="Whether an expected-response correction was successfully logged.",
-    )
-
-
-# ---------------------------------------------------------------------------
-# Dataset + evaluation result + run comparison schemas
-# ---------------------------------------------------------------------------
-
-
-class DatasetResponse(BaseModel):
-    """Metadata for a registered dataset."""
-
-    id: str = Field(description="Unique dataset identifier.")
-    name: str = Field(description="Human-readable dataset name.")
-    row_count: int = Field(description="Number of rows/examples in the dataset.")
-    format: str = Field(description="File format (json or jsonl).")
-    module_slug: str | None = Field(
-        default=None, description="Associated module slug, when provided."
-    )
-    created_at: str = Field(description="ISO-8601 creation timestamp.")
-
-
-class DatasetListResponse(BaseModel):
-    """Paginated dataset listing."""
-
-    items: list[DatasetResponse] = Field(description="Dataset list items.")
-    total: int = Field(description="Total matching datasets.")
-    offset: int = Field(description="Current pagination offset.")
-    limit: int = Field(description="Current page size.")
-    has_more: bool = Field(description="Whether more results exist beyond this page.")
-
-
-class DatasetDetailResponse(DatasetResponse):
-    """Dataset metadata with sample rows and URI."""
-
-    sample_rows: list[dict[str, Any]] = Field(
-        description="First rows from the dataset as preview."
-    )
-    uri: str = Field(description="Filesystem path to the dataset file.")
-
-
-class EvaluationResultItem(BaseModel):
-    """A single per-example evaluation result."""
-
-    id: str = Field(description="Unique evaluation result identifier.")
-    example_index: int = Field(description="Zero-based index in the dataset.")
-    input_data: str = Field(description="JSON-serialized input fields.")
-    expected_output: str | None = Field(
-        default=None, description="Expected/gold output."
-    )
-    predicted_output: str | None = Field(
-        default=None, description="Model predicted output."
-    )
-    score: float = Field(description="Score for this example (0.0-1.0).")
-
-
-class EvaluationResultsResponse(BaseModel):
-    """Paginated evaluation results for a run."""
-
-    items: list[EvaluationResultItem] = Field(description="Evaluation result items.")
-    total: int = Field(description="Total evaluation results for the run.")
-    offset: int = Field(description="Current pagination offset.")
-    limit: int = Field(description="Current page size.")
-    has_more: bool = Field(description="Whether more results exist beyond this page.")
-
-
-class PromptSnapshotItem(BaseModel):
-    """A before or after prompt snapshot for a predictor."""
-
-    predictor_name: str = Field(description="Predictor name from named_predictors().")
-    prompt_type: str = Field(description="Snapshot type: 'before' or 'after'.")
-    prompt_text: str = Field(description="Full prompt/instruction text.")
-
-
-class RunComparisonItem(BaseModel):
-    """Summary of a single run for cross-run comparison."""
-
-    run_id: str = Field(description="Optimization run identifier.")
-    program_spec: str = Field(description="DSPy program specification optimized.")
-    validation_score: float | None = Field(
-        default=None, description="Validation score from the run."
-    )
-    prompt_snapshots: list[PromptSnapshotItem] = Field(
-        description="Before/after prompt snapshots for this run."
-    )
-
-
-class RunComparisonResponse(BaseModel):
-    """Cross-run comparison payload."""
-
-    runs: list[RunComparisonItem] = Field(description="Compared run summaries.")
-
-
-class SandboxListItem(BaseModel):
-    """Single sandbox entry returned by the sandbox list endpoint."""
-
-    id: str = Field(description="Sandbox identifier.")
-    name: str = Field(description="Sandbox name.")
-    state: str = Field(description="Sandbox state (e.g. started, stopped, archived).")
-    created_at: str | None = Field(
-        default=None, description="ISO-8601 creation timestamp when available."
-    )
-    volume_name: str | None = Field(
-        default=None,
-        description="Name of the persistent volume attached to the sandbox.",
-    )
-    labels: dict[str, str] = Field(
-        default_factory=dict, description="Custom labels attached to the sandbox."
-    )
-    cpu: int | None = Field(default=None, description="Allocated CPU cores.")
-    memory: int | None = Field(default=None, description="Allocated memory in GiB.")
-    disk: int | None = Field(default=None, description="Allocated disk in GiB.")
-
-
-class SandboxDetailResponse(BaseModel):
-    """Detailed response for a single sandbox."""
-
-    id: str = Field(description="Sandbox identifier.")
-    name: str = Field(description="Sandbox name.")
-    state: str = Field(description="Sandbox state (e.g. started, stopped, archived).")
-    created_at: str | None = Field(
-        default=None, description="ISO-8601 creation timestamp when available."
-    )
-    volume_name: str | None = Field(
-        default=None,
-        description="Name of the persistent volume attached to the sandbox.",
-    )
-    labels: dict[str, str] = Field(
-        default_factory=dict, description="Custom labels attached to the sandbox."
-    )
-    cpu: int | None = Field(default=None, description="Allocated CPU cores.")
-    memory: int | None = Field(default=None, description="Allocated memory in GiB.")
-    disk: int | None = Field(default=None, description="Allocated disk in GiB.")
-    env_vars: dict[str, str] = Field(
-        default_factory=dict,
-        description="Environment variables configured for the sandbox.",
-    )
-    image: str | None = Field(
-        default=None, description="Base image or declarative image used by the sandbox."
-    )
-    snapshot: str | None = Field(
-        default=None, description="Snapshot name used to create the sandbox."
-    )
-    language: str | None = Field(
-        default=None, description="Programming language of the sandbox."
-    )
-    auto_stop_interval: int | None = Field(
-        default=None,
-        description="Minutes of inactivity before auto-stopping.",
-    )
-    auto_archive_interval: int | None = Field(
-        default=None,
-        description="Minutes after stop before archiving to cold storage.",
-    )
-    auto_delete_interval: int | None = Field(
-        default=None,
-        description="Minutes after archive before permanent deletion.",
-    )
-    ephemeral: bool | None = Field(
-        default=None, description="Whether the sandbox is ephemeral."
-    )
-    network_block_all: bool | None = Field(
-        default=None, description="Whether all outbound network is blocked."
-    )
-    network_allow_list: str | None = Field(
-        default=None, description="Comma-separated list of allowed domains."
-    )
-    volumes: list[dict[str, Any]] = Field(
-        default_factory=list, description="Detailed volume mounts."
-    )
-
-
-class SandboxListResponse(BaseModel):
-    """Response for the sandbox list endpoint."""
-
-    items: list[SandboxListItem] = Field(
-        default_factory=list, description="Available sandboxes."
-    )
-    total: int = Field(description="Total number of sandboxes.")
-    page: int = Field(default=1, description="Current page number.")
-    total_pages: int = Field(default=1, description="Total number of pages.")
-
-
-class SandboxArchiveResponse(BaseModel):
-    """Result payload after archiving a sandbox."""
-
-    ok: bool = Field(
-        default=True,
-        description="Whether the sandbox was archived successfully.",
-    )
-
-
-class RunStepItem(BaseModel):
-    """Single execution step for a run."""
-
-    id: str = Field(description="Durable step identifier.")
-    step_index: int = Field(description="Step position within the run.")
-    step_type: str = Field(description="Step type (e.g. tool_call, reasoning).")
-    tool_name: str | None = Field(
-        default=None, description="Tool name when applicable."
-    )
-    tokens_in: int | None = Field(default=None, description="Input token count.")
-    tokens_out: int | None = Field(default=None, description="Output token count.")
-    latency_ms: int | None = Field(
-        default=None, description="Step latency in milliseconds."
-    )
-    created_at: str = Field(description="ISO-8601 creation timestamp.")
-
-
-class RunStepListResponse(BaseModel):
-    """Paginated execution step list for a run."""
-
-    items: list[RunStepItem] = Field(description="Step list items.")
-    total: int = Field(description="Total steps in run.")
-    offset: int = Field(description="Current pagination offset.")
-    limit: int = Field(description="Current page size.")
-    has_more: bool = Field(description="Whether more steps exist beyond this page.")
-
-
-class SessionExportRequest(BaseModel):
-    """Request body for exporting a session's turns as a GEPA training dataset."""
-
-    module_slug: str = Field(
-        description="Target GEPA module slug whose dataset keys "
-        "determine the export column mapping."
-    )
-
-
-class TranscriptTurnInput(BaseModel):
-    """Single transcript turn used to build a GEPA dataset."""
-
-    user_message: str | None = Field(
-        default=None,
-        description="User prompt/content for the turn.",
-    )
-    assistant_message: str | None = Field(
-        default=None,
-        description="Assistant response/content for the turn.",
-    )
-
-
-class TranscriptDatasetRequest(BaseModel):
-    """Request body for converting transcript turns into a GEPA dataset."""
-
-    module_slug: str = Field(
-        description="Target GEPA module slug whose dataset keys determine row mapping."
-    )
-    title: str | None = Field(
-        default=None,
-        description="Optional human-readable transcript title used for dataset naming.",
-    )
-    turns: list[TranscriptTurnInput] = Field(
-        description="Transcript turns to convert into dataset rows.",
-    )
-
-
-class MemoryItemResponse(BaseModel):
-    """Single memory item returned by the memory browse endpoint."""
-
-    id: str = Field(description="Durable memory item identifier.")
-    scope: str = Field(
-        description="Memory scope (e.g. user, tenant, workspace, run, session)."
-    )
-    scope_id: str = Field(description="Identifier within the scope.")
-    kind: str = Field(description="Memory kind (e.g. fact, observation, preference).")
-    source: str = Field(description="Memory source (e.g. user, agent, system).")
-    status: str = Field(description="Memory status (e.g. active, archived).")
-    content_text: str | None = Field(
-        default=None, description="Textual content when available."
-    )
-    importance: int = Field(description="Importance score (0-100).")
-    tags: list[str] = Field(default_factory=list, description="Associated tags.")
-    created_at: str = Field(description="ISO-8601 creation timestamp.")
-
-
-class MemoryListResponse(BaseModel):
-    """Paginated memory item list response."""
-
-    items: list[MemoryItemResponse] = Field(description="Memory item list items.")
-    total: int = Field(description="Total matching memory items.")
-    offset: int = Field(description="Current pagination offset.")
-    limit: int = Field(description="Current page size.")
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from .base import *  # noqa: F403
+    from .feedback import *  # noqa: F403
+    from .memory import *  # noqa: F403
+    from .optimization import *  # noqa: F403
+    from .runtime import *  # noqa: F403
+    from .sandbox import *  # noqa: F403
+    from .sessions import *  # noqa: F403
+    from .volumes import *  # noqa: F403
+    from .websocket import *  # noqa: F403
+
+# Runtime lazy re-exports to preserve import compatibility
+from .base import *  # noqa: F403
+from .feedback import *  # noqa: F403
+from .memory import *  # noqa: F403
+from .optimization import *  # noqa: F403
+from .runtime import *  # noqa: F403
+from .sandbox import *  # noqa: F403
+from .sessions import *  # noqa: F403
+from .volumes import *  # noqa: F403
+from .websocket import *  # noqa: F403
